Support OpenPAM
fixes: #24, #30
